Folders added
Added folder structure to repo
