+++ include/Segment.h
