+++ src/Segment.cpp
